test: size 2 osu and kripke
I did my best to run both with caliper and without.
I am not sure the osu trace is working perfectly, so if
ultimately we are missing data for it, so be it.
I am running out of time to do a lot of work and if
I get hung up on this I will not be able to do it.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>